Added Screenshot, updated information in Style.css
I think a number of changes have happened with the style.css file.  What I remember is:

Updated the tags, only mentioning what it implements, as well as the parent theme.
Updated the version number.
May have made changes to description, but I don't think so.

I added the screenshot.png file, which should allow the display of the theme.  This I think is current to version 0.0.2 of the theme.

So far those are the changes I have made to the theme with this update.