Updated Version Number
There seems to be no need to worry about any issues with this version any more.  Time to move on to the next version.  Which will be 0.0.4.

We should be good to go before too long.  I am wondering if it is getting close to the time to upgrade to 0.1.0 which would be the first "stable" branch, and to 0.2.0 which would be the second development branch.