Read functions.php.  Add woocommerce and shortcodes.  Move customizer.
Changes to get this to be in line with how I want things for the child theme.  This gets the functions.php back to where it should have been...

customizer.php, shortcodes.php, and woocommerce.php have all been put where we want them.

Shortcodes and woocommerce are empty files.